sli: check cgroup's proactive monitoring statistics and report the event if necessary
If there is not any task of cgroup running on the cpu, there is
no necessary to do the monitoring event check for cgroup. So we
put the checkpoint in the tick process to reduce the overhead
of system(if tasks of cgroup are not running, the tick interrupt
will not hit the task).

Signed-off-by: Bin Lai <robinlai@tencent.com>