sli: remove the page_alloc trace item
Page_alloc is may be a hot spot in the system, the overhead of
trace will increase with the number of page alloc operations.
And we only focus on the page alloc latency caused by memory
compaction or reclaim. So we remove the useless page alloc item
to reduce the overhead of trace.

Signed-off-by: Bin Lai <robinlai@tencent.com>