sched: make cfs scheduler support cpumask boosting
No matter periodic load balance or wakeup load blance, the scheduler
have to comply with tasks's cpu affinity when the scheduler select
the suitable cpu. If we want to select the cpu that outside the
affinity cpumask of tasks, we have to change the cpu affinity of
tasks. The cpumask modification is time comsuming, the frequent
cpumask modification could lead to performance regress(and the
order of task's pi_lock and rq lock is difficult to handle). So
we add this patch to allow the scheduler pick the task to the cpu
that outside its own cpumask.

Signed-off-by: Bin Lai <robinlai@tencent.com>